requirements.txt
# Core dependencies
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
Pillow==10.2.0
numpy==1.26.4
opencv-python==4.9.0.80
pytesseract==0.3.10
reportlab==4.0.4
ultralytics==8.1.0
torch==2.2.0
torchvision==0.17.0

# Additional utilities
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.0 